requirements.txt
attrs
numpy
Pillow
pytest
tqdm
